lib_openai_agent/requirements.txt
openai-agents>=0.0.16
langfuse>=3.0.0
nest-asyncio>=1.6.0
pydantic-ai[logfire]>=0.2.16
annotated-types>=0.7.0
anyio>=4.6.0
click>=8.1.7
fastapi>=0.115.0
h11>=0.14.0
idna>=3.10
pydantic>=2.10.0
pydantic_core>=2.23.4
sniffio>=1.3.1
starlette>=0.38.6
typing_extensions>=4.12.2
uvicorn>=0.31.0
python-dotenv
openai>=1.81.0
redis>=5.2.0
requests>=2.32.3
asyncio>=3.4.3
beautifulsoup4>=4.12.3
bleach>=6.2.0
pytz>=2025.2
slowapi>=0.1.9
gunicorn>=23.0.0
colorama>=0.4.6